add scrim to side menu, split mapbox-navigation into two components
